Make items more easily clickable
Currently, the title in compact mode is an inline tag, which means that if it starts wrapping to a new line then any white space isn't clickable to expand the article. These changes make it more easily clickable, don't interfere with the "open" and "favourite" buttons, make wrapped titles align more neatly, and prevent the "time ago" text being a dead spot.