Improve article enhancer for sueddeutsche.de
This improves the article enhancer for sueedeutsche.de to fetch just the
summary from the top of each article instead of the whole article which
bypasses the paywall sueddeutsche.de introduced last week (and that is
not the purpose of the news app, right?). The summary is more detailed
than the one delivered via RSS so the enhancer is still usefull. Plus it
adds the header image of the article.