Always show the unread articles.
The unread articles folder will always be visible, regardless
of the "Show all" setting. If that setting is on, then an
additional "All articles" folder will be show as before, but
it doesn't substitute the "Uread articles" one.

A new URL /apps/news/#/items/unread is also created that jumps
to the unread articles.

Signed-off-by: Cesar Enrique Garcia Dabo <cquike@arcor.de>