Make the "open" keyboard shortcut work faster
Previously when pressing the `O` key on article list, the handler for
that keypress first simulated a click on that event in order to mark
it as read, and only then opened the website that item links to in
another tab. When having a lot of items on screen this caused a huge
delay between pressing `O` and opening the linked article in a new
tab. The delay was sometimes 5, even 10 whole seconds. This simple fix
makes it so the article opens first, and then the click simulation
happens afterwards.

Signed-off-by: Kuba Orlik <kontakt@kuba-orlik.name>
Signed-off-by: Benjamin Brahmer <info@b-brahmer.de>