fix(ui): "a.data.items is undefined" on first run
When enabling the app for the very first time and opening the view, a
toast notification pops up with the following message:

> TypeError: a.data.items is undefined

This is caused by a silenced exception in lib/Controller/ItemController.
Due to it, not a single key is set in the $return array.

Responses returned from such a controller will be serialized as JSON.
However, the returned value is no longer a key-value array, but an
(empty) list and the serializer will convert it to `[]` instead of `{}`
what it should've been.

Now it will return a dummy response containing all the keys but
empty values, causing the serializer to return an object.

Signed-off-by: Vincent Neubauer <v.neubauer@darlor.de>